592 - at most one of state/state_name can be set
This extends the validation of state info in addresses to ensure that
only one of {state,state_name} is set in a valid address, and all other
conditions are invalid. However, if :address_requires_state is false,
then there should be no state information in the record.

The condition for only one to be set introduces a conflict when existing
records are changed from state to state_name. There's a pre-validation
filter which blanks the previous information to allow the new
information to be saved; eg if the address had a state, then adding
state_name info will cause the state to be dropped and the state_name
to be kept.

Detailed tests have been added.

[592 state:patched]